# Reducing memory usage while evaluating reflective EL expressions