@ChristopherSchultz 2024-10-22
I was also surprised recently to discover that the JVM does not cache results of `Class.getMethods` and similar calls. Maybe it's returning a clone each time, but either way it takes time and generates garbage.

So there is a (positive) GC impact of this change as well.