@markt-asf 2024-10-30
I'd also like to see the key remove from the cache when the value expires.