@ChristopherSchultz 2024-10-30
> I'd also like to see the key remove from the cache when the value expires.

+1

So I think your technique @jengebr is good, but the cache just needs to be pulled-up into the application scope or similar. Definitely don't cache directly in the Util class. It probably needs to be in a class that is loaded into the application's ClassLoader to be as clean as possible.

Actually, I think we *must not* change the `javax.el.Util` class, right?