@jengebr 2024-10-30
Thanks, I get your points about caching scope and cleanup.

@ChristopherSchultz what prevents changes to `javax.el.Util`?