@jengebr 2024-10-30
This complexity led me to try a fallback plan, which can be faster: if the method has zero parameters, retrieve the Method by name.  This works because there is no confusion about parameter types.  Example:

```
        if (paramTypes.length == 0) {
            try {
                Method method = clazz.getMethod(methodName, paramTypes);
                return method;
            } catch (NoSuchMethodException | SecurityException e) {
                // fall through to broader, slower logic
            }
        }
```

Test data shows this is 2x faster than caching for the fast case, and has no appreciable impact on the others.  That's an excellent outcome for my application, not sure how you feel about the general case?