@ChristopherSchultz 2024-10-30
> @ChristopherSchultz what prevents changes to `javax.el.Util`?

That class doesn't "belong" to us. Depending upon how it changes, we might fail the TCK. It's just better to put it in something under `org.apache.whatever`.