@markt-asf 2024-10-30
I'm fine with the zero-arg optimisation. If others want the more general caching solution it can still be implemented.