@markt-asf 2024-10-31
I'm planning on applying this manually to main and then back-porting. So far I've made the following changes:
- refactored the inner test classes to align with the rest of the package
- ignored the caching code that wasn't (fully) removed
- Called Util.getMethod() before returning to address some test failures caused by visibility issues

Just running the full test suite before committing to check I haven't missed anything.