@markt-asf 2024-10-31
Those were the only changes required. Fix applied to main, 11.0.1, 10.1.x and 9.0.x.